+++ demisto_sdk/commands/common/constants.py
@@ -32,9 +32,9 @@ RELEASE_NOTES_DIR = 'ReleaseNotes'
 TESTS_DIR = 'Tests'
 DOC_FILES_DIR = 'doc_files'
 DOCUMENTATION_DIR = 'Documentation'
-OBJECT_MODULE_DIR = 'ObjectModules'
-OBJECT_TYPE_DIR = 'ObjectTypes'
-OBJECT_FIELD_DIR = 'ObjectFields'
+GENERIC_MODULE_DIR = 'genericModules'
+GENERIC_TYPE_DIR = 'genericTypes'
+GENERIC_FIELD_DIR = 'genericFields'
 
 SCRIPT = 'script'
 AUTOMATION = 'automation'
@@ -97,9 +97,9 @@ class FileType(Enum):
     LANDING_PAGE_SECTIONS_JSON = 'landingPage_sections.json'
     CONTRIBUTORS = 'contributors'
     XSOAR_CONFIG = 'xsoar_config'
-    OBJECT_TYPE = 'ObjectType'
-    OBJECT_FIELD = 'ObjectField'
-    OBJECT_MODULE = 'ObjectModules'
+    GENERIC_TYPE = 'genericType'
+    GENERIC_FIELD = 'genericField'
+    GENERIC_MODULE = 'genericModules'
 
 
 RN_HEADER_BY_FILE_TYPE = {