+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 * Fixed an issue where a redundant print occurred on release notes validation.
 * Added new validation in the **validate** command to ensure that the minimal fromVersion in a widget of type metrics will be 6.2.0.
 * Added the *--release-notes* flag to demisto-sdk to get the current version release notes entries.
-* Updated the **create-id-set** command to support objects modules, objects fields and objects types.
+* Updated the **create-id-set** command to support generic modules, generic fields and generic types.
 # 1.4.2
 * Added to `pylint` summary an indication if a test was skipped.
 * Added to the **init** command the option to specify fromversion.